Significant update
Implementation of the betfair Starting prices for each dog, increases performance on profit across the board by ~12%. Still some kinks to work out, mainly why the amount of bfsp missing is signifcantly larger for winning dogs. Potentially an issue with formatting or matching in the DF join.

Implemented a custom loss function adapted from MSE, calculates MSE for first 3 place dogs only. Implemented a hyperparameter sweep for loss functions to determine which performs best when optimizing for profit not for accuracy (not always the same).

Various other fixes

Implementation of the betfair Starting prices for each dog, increases performance on profit across the board by ~12%. Still some kinks to work out, mainly why the amount of bfsp missing is signifcantly larger for winning dogs. Potentially an issue with formatting or matching in the DF join.

Implemented a custom loss function adapted from MSE, calculates MSE for first 3 place dogs only. Implemented a hyperparameter sweep for loss functions to determine which performs best when optimizing for profit not for accuracy (not always the same).

Various other fixes